rag_chatbot/requirements.txt
langchain
langchain-groq
langchain-community
langchain-chroma
langchain-text-splitters
chromadb
sentence-transformers
streamlit
python-dotenv
pymupdf